+++ CesiumGeospatial/include/CesiumGeospatial/S2CellID.h
@@ -4,11 +4,13 @@
 #include "GlobeRectangle.h"
 #include "Library.h"
 
-#include <CesiumGeometry/QuadtreeTileID.h>
-
 #include <array>
 #include <cstdint>
 #include <string_view>
+
+namespace CesiumGeometry {
+struct QuadtreeTileID;
+}
 
 namespace CesiumGeospatial {