+++ CesiumGeospatial/src/Projection.cpp
@@ -78,6 +78,33 @@ GlobeRectangle unprojectRectangleSimple(
   return std::visit(Operation{rectangle}, projection);
 }
 
+CesiumGeometry::AxisAlignedBox projectRegionSimple(
+    const Projection& projection,
+    const BoundingRegion& region) {
+  CesiumGeometry::Rectangle rectangle =
+      projectRectangleSimple(projection, region.getRectangle());
+  return CesiumGeometry::AxisAlignedBox(
+      rectangle.minimumX,
+      rectangle.minimumY,
+      region.getMinimumHeight(),
+      rectangle.maximumX,
+      rectangle.maximumY,
+      region.getMaximumHeight());
+}
+
+BoundingRegion unprojectRegionSimple(
+    const Projection& projection,
+    const CesiumGeometry::AxisAlignedBox& box) {
+  GlobeRectangle rectangle = unprojectRectangleSimple(
+      projection,
+      CesiumGeometry::Rectangle(
+          box.minimumX,
+          box.minimumY,
+          box.maximumX,
+          box.maximumY));
+  return BoundingRegion(rectangle, box.minimumZ, box.maximumZ);
+}
+
 glm::dvec2 computeProjectedRectangleSize(
     const Projection& projection,
     const CesiumGeometry::Rectangle& rectangle,