+++ CesiumGltfReader/src/CesiumGltfReader/decodeDraco.cpp
@@ -154,6 +154,7 @@ void copyDecodedIndices(
   indicesBufferView.target =
       CesiumGltf::BufferView::Target::ELEMENT_ARRAY_BUFFER;
   pIndicesAccessor->type = CesiumGltf::Accessor::Type::SCALAR;
+  pIndicesAccessor->byteOffset = 0;
 
   static_assert(sizeof(draco::PointIndex) == sizeof(uint32_t));
 
@@ -209,9 +210,9 @@ void copyDecodedAttribute(
   CESIUM_TRACE("CesiumGltfReader::copyDecodedAttribute");
   CesiumGltf::Model& model = readModel.model.value();
 
-  if (pAccessor->count > pMesh->num_points()) {
-    readModel.warnings.emplace_back("There are fewer decoded Draco indices "
-                                    "than are expected by the accessor.");
+  if (pAccessor->count != pMesh->num_points()) {
+    readModel.warnings.emplace_back("Attribute accessor.count doesn't match "
+                                    "with number of decoded Draco vertices.");
 
     pAccessor->count = pMesh->num_points();
   }