+++ CHANGES.md
@@ -23,6 +23,11 @@
 - Added `GltfContent::applyRtcCenter` and `applyGltfUpAxisTransform`.
 - Clipping polygon edges now remain sharp even when zooming in past the available geometry detail.
 
+##### Fixes :wrench:
+
+- Fixed a bug that could cause memory corruption when a decoded Draco mesh was larger than indicated by the corresponding glTF accessor.
+- Fixed a bug that could cause the wrong triangle indices to be used for a Draco-encoded glTF.
+
 ### v0.9.0 - 2021-11-01
 
 ##### Breaking Changes :mega: